+++ models/feed.js
@@ -25,6 +25,10 @@ Feed.init(
        user_id: {
         type: DataTypes.INTEGER,
         allowNull: false,
+        references: {
+          model: "user",
+          key: "id", 
+        },
         },
        //user posts
        content: {
